requirements.txt
Django==2.2
djangorestframework==3.11.0
gunicorn==19.7.1
psycopg2==2.8.5
PyJWT==1.5.0
drf-extensions==0.6.0

